Added the apiservice.fundStore.
The fundStore adapts the domain.FundRepository iterface to the
json api specificiation. This commit adds the domain interfaces
Fund and FundRepository.

The implementation of the CRUD methods for a fundStore is not yet
complete. The List() method is implemented but the others will
panic when called.

The implementation is the CurrencyName() methon in domain/currency.go
is a stub to make the tests pass.