Updated assumed configuration and added @cleandb
Updated the api prefix and the dsn configuration to match what
was implemented or what can be assumed.

Added the @cleandb tag and a before handler for the tag to ensure
a clean database before any scenarios to which the tag is applied. The
handler currently drops and recreates the database named in the assumed
dsn.