Extened to step defitions to cover another scenario.
Refactored the step defitions out of the init() function into their own
functions to facilitate reusing one of them for a step needed for the
"Bookkeeper adds a second fund" scenario.

Revised the step function for checking for the exisitance of a fund to
be more robust in the face of listing multiple funds.